// mex/editor/client/src/app/shared/models/generated/contact-point.ts
export const AdditiveContactPointSchema = z.object({
  email: z.array(z.string().regex(new RegExp(emailPattern))).optional(),
  $type: z.literal("AdditiveContactPoint"),
  supersededBy: z.string().regex(new RegExp(identifierPattern)).nullish(),
});

/**
 * TypeScript type inferred from AdditiveContactPointSchema.
 */
export type AdditiveContactPoint = z.infer<typeof AdditiveContactPointSchema>;