// mex/editor/client/src/app/shared/models/generated/person.ts
export const AdditivePersonSchema = z.object({
  affiliation: z.array(z.string().regex(new RegExp(identifierPattern))).optional(),
  email: z.array(z.string().regex(new RegExp(emailPattern))).optional(),
  familyName: z.array(z.string()).optional(),
  fullName: z.array(z.string()).optional(),
  givenName: z.array(z.string()).optional(),
  isniId: z.array(z.string().regex(new RegExp(isniIdPattern))).optional(),
  memberOf: z.array(z.string().regex(new RegExp(identifierPattern))).optional(),
  orcidId: z.array(z.string().regex(new RegExp(orcidIdPattern))).optional(),
  $type: z.literal("AdditivePerson"),
  supersededBy: z.string().regex(new RegExp(identifierPattern)).nullish(),
});

/**
 * TypeScript type inferred from AdditivePersonSchema.
 */
export type AdditivePerson = z.infer<typeof AdditivePersonSchema>;